Hacky-dacky solution
We monitor whether or not the search bar has focus.
We support two cases:

1) press enter, which triggers the following events:

  - search-button click
  - searchBar blur

2) click search-button, which triggers the following events:

  - search-button mousedown
  - searchBar blur
  - search-button click

In the latter case we need to monitor mousedown and save the state the searchBar is in when this happens.
Finally, on click, if EITHER the search bar is currently in focus OR the search bar WAS in focus at last mousedown, trigger a search.
Otherwise, focus the search bar as usual