Blog as collection
This commit changes blog posts into a Jekyll collection, relying on some standardized fields and the existence of index.html and aside.html in each dir.
This gives us better separation of concerns (aside and content), and makes it easier to add posts in the future, as some fields are now generated from the same values, rather than entered in duplicate.